Renamed configuration package to config
