+++ cmd/mining/main.go
@@ -3,22 +3,20 @@ package main
 import (
 	log "github.com/sirupsen/logrus"
 	"rampart/internal/mining"
-	"rampart/internal/mining/configs"
+	"rampart/internal/mining/config"
 	"rampart/internal/mining/domria"
 )
 
 func main() {
 	log.SetLevel(log.DebugLevel)
 	log.Debug("main: mining started")
-	config, err := configs.NewMining()
-	exitIfError(err)
-	prospector := domria.NewProspector(mining.Secondary, config.Prospectors.Domria)
-	exitIfError(prospector.Prospect())
-	log.Debug("main: mining finished")
-}
-
-func exitIfError(err error) {
+	cfg, err := config.NewMining()
 	if err != nil {
 		log.Fatal(err)
 	}
+	prospector := domria.NewProspector(mining.Secondary, cfg.Prospectors.Domria)
+	if err = prospector.Prospect(); err != nil {
+		log.Fatal(err)
+	}
+	log.Debug("main: mining finished")
 }
