+++ internal/mining/config/fetcher.go
@@ -1,4 +1,4 @@
-package configs
+package config
 
 import (
 	"fmt"
