+++ internal/mining/config/geocoder.go
@@ -1,4 +1,4 @@
-package configs
+package config
 
 import (
 	"fmt"
