+++ internal/mining/domria/validator.go
@@ -3,10 +3,10 @@ package domria
 import (
 	log "github.com/sirupsen/logrus"
 	"rampart/internal/mining"
-	"rampart/internal/mining/configs"
+	"rampart/internal/mining/config"
 )
 
-func newValidator(config *configs.Validator) *validator {
+func newValidator(config *config.Validator) *validator {
 	return &validator{
 		config.MinPrice,
 		config.MinTotalArea,
